fixed test file upload, deploy is halfway there

+++ helper.py
@@ -1,3 +1,4 @@
+import _csv
 import shutil
 from abc import ABCMeta, abstractmethod
 
@@ -255,15 +256,14 @@ class Tabular(Helper):
 
     def test_upload(self, request):
         test_file = get_filename(request)
+        test_filename = ''
         try:
-
             test_filename = os.path.join(self._dataset.get_base_path(), 'test', test_file)
             df_test = sys_ops.bytestr2df(request.get_json()['file'], test_filename)
-            print('ok')
             sys_ops.check_df(df_test, self._dataset.get_df(), self._dataset.get_targets(), test_filename)
-            print('pos aqui no ha sio')
-        except ValueError:
-            os.remove(test_filename)
+        except (ValueError, _csv.Error):
+            if os.path.isfile(test_filename):
+                os.remove(test_filename)
             return "The file contents are not valid."
         return "ok"
